Split the config helpers into two pieces
The Config class is useful for more than just carbon.conf;
storage-schemas.conf uses the same format.
